refactor(db): extract raw pool to break the connection/slowQueryLog cycle
connection.ts statically imported recordSlowQuery on the hot query path
while slowQueryLog lazily imported getDatabasePool back from connection —
a documented import cycle. Move the pool singleton, config, and lifecycle
into rawPool.ts; connection.ts re-exports them so consumers are
unaffected, and slowQueryLog imports the pool from the leaf module. Both
import/no-cycle disables removed. No behavior change.
